refactor(pi-process): move Pi process execution subsystem out of runner.py (Issue #300)
Move the 11 Pi-process functions (plus _refresh_session_evidence, added by
Issue #656) into a self-contained src/orbi/pi_process.py along with the
Pi-process tuning constants and failure classes it owns. runner.py now
imports stream_pi/_log_provider_config_loaded and the constants/exceptions
from pi_process; the two runner-side lifecycle hooks stream_pi needs
(issue_context, set_active_pi) are imported lazily to avoid a circular
import. No logic changes.
